Change the branding technique for string types.
If you use the old brands with a library like zod, it will try to
replicate the type definition structurally rather than using the name
'StringUserID', and it won't let you import the secret (even if you
try to).

The new branding technique allows the structure of the type to be
replicated without any issues.